Update sumo.md
SUMO is typed.  The class hierarchy is the type structure.  All relations have a defined type signature specified with the domain, domainSubclass (and for functions range and rangeSubclass statements).  Enache's comment is not accurate.  We regularly use the TPTP translation of SUMO with Vampire and Eprover to check for the logical consistency of the axioms.  That includes (but is not limited to) checking the type constraints on relations and their arguments.